moving the fixtures so can be use acorss other class
